carcluate함수 분리
