+++ src/basic/hooks/useCart.ts
@@ -24,11 +24,15 @@
 // - getRemainingStock: 재고 확인 함수
 // - clearCart: 장바구니 비우기 함수
 
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 
 import { CartItem } from '../../types';
+import { ProductWithUI } from '../App';
+import { getRemainingStock } from '../utils/calculateItem';
 
-export function useCart() {
+export function useCart(
+  addNotification: (message: string, type?: 'error' | 'success' | 'warning') => void
+) {
   const [cart, setCart] = useState<CartItem[]>(() => {
     const saved = localStorage.getItem('cart');
     if (saved) {
@@ -41,6 +45,38 @@ export function useCart() {
     return [];
   });
   const [totalItemCount, setTotalItemCount] = useState(0);
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product, cart);
+      if (remainingStock <= 0) {
+        addNotification('재고가 부족합니다!', 'error');
+        return;
+      }
+
+      setCart((prevCart) => {
+        const existingItem = prevCart.find((item) => item.product.id === product.id);
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id ? { ...item, quantity: newQuantity } : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification('장바구니에 담았습니다', 'success');
+    },
+    [cart, addNotification, getRemainingStock]
+  );
 
   // 카트 아이템 수 계산
   useEffect(() => {
@@ -57,34 +93,5 @@ export function useCart() {
     }
   }, [cart]);
 
-  // const removeFromCart = useCallback((productId: string) => {
-  //   setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
-  // }, []);
-
-  // const updateQuantity = useCallback(
-  //   (productId: string, newQuantity: number) => {
-  //     if (newQuantity <= 0) {
-  //       removeFromCart(productId);
-  //       return;
-  //     }
-
-  //     const product = products.find((p) => p.id === productId);
-  //     if (!product) return;
-
-  //     const maxStock = product.stock;
-  //     if (newQuantity > maxStock) {
-  //       addNotification(`재고는 ${maxStock}개까지만 있습니다.`, 'error');
-  //       return;
-  //     }
-
-  //     setCart((prevCart) =>
-  //       prevCart.map((item) =>
-  //         item.product.id === productId ? { ...item, quantity: newQuantity } : item
-  //       )
-  //     );
-  //   },
-  //   [products, removeFromCart, addNotification, getRemainingStock]
-  // );
-
-  return { cart, setCart, totalItemCount };
+  return { cart, setCart, totalItemCount, addToCart };
 }
